Adding FeedForward and testing notebook
# Adding FeedForward network and a notebook for testing

FeedForward.py is a simple structure for a feedforward neural network. It hasn't been deeply tested, but for the minimal example of addition of two numbers, it seems to work well enough. A demonstration of this (for addition of two numbers) is given in the Jupyter notebook Testing_FF_Network.ipynb, which also includes some more faffing around (which is notably less successful).